+++ Anytype/Sources/ServiceLayer/Smartblock/SubscriptionService/SubscriptionIds.swift
@@ -6,9 +6,9 @@ struct SetSubsriptionData: Hashable {
     let sorts: [DataviewSort]
     let filters: [DataviewFilter]
     let relations: [DataviewViewRelation]
-    let currentPage: Int
+    let currentPage: Int64
     
-    init(dataView: BlockDataview, view: DataviewView, currentPage: Int) {
+    init(dataView: BlockDataview, view: DataviewView, currentPage: Int64) {
         self.source = dataView.source
         self.sorts = view.sorts
         self.filters = view.filters
